feat: extract tests utils
